CompilerGenerator

Description

This module uses the ParserGenerator module, and as such, the .gmr file input for this module follows the same structure as that. Similarly to the ParserGenerator module, this module can be invoked both within haskell in a pure computation, or by command line/via an IO computation. The IO method of invoking the compiler generator will handle all input and output files for you, whereas using the haskell functions is a little more involved, as you are expected to correctly name the file and generate the requirements.

Documentation

runCompilerGenerator :: String -> IO ()

This function takes a path to file with no extension, and expects a .gmr and .smt file be present. For example, calling this function with examples/mt will expect that examples/mt.gmr and examples/mt.smt exist. The 4 files required for a compiler will be generated in the same location as the input files, named after the input files. For example, a compiler by the name mt would generate the following files:

mtcompiler.hs

mtparser.hs

mtsemantics.hs

parserrequirements.hs

generateCompiler

Arguments

:: String
Contents of the .gmr file

-> String
Contents of the .smt file

-> String
Compiler name

-> Result (String, String, String)
Failure-capable tuple of (parserCode, semanticsCode, compilerCode)

Generates the code for the parser, semantics and compiler from the input definitions. This function can fail for invalid inputs.

generateSemantics

Arguments

:: String
Contents of the .smt file

-> String
Module name for the parser

-> String
Compiler name

-> Result (String, String, Bool)
Failure-capable tuple of (semanticsCode, fileExtension, supportsIncludesFlag)

Generates the semantics code from an .smt file, along with information needed to generate the main compiler file.