MainCodeGenerator

Description

The code generated by this file handles all the IO actions for the generate compiler, and acts as the entry point. This code will differ greatly depending on whether or not includes are supported, since the majority of the handling for this feature is within this file. The IO actions supported by this file include argument and flag handling for command-line use, input and output file path specification, and verbose command-line output.

Documentation

generateMainCode

Arguments

:: String
Parser module name

-> String
Semantics module name

-> String
Input file extension

-> Bool
Whether or not the generated compiler supports includes

-> String
The main compiler code

This function returns the entire compiler main file, and cannot fail.