ParserRequirements

Description

This module is copied to any generated parser and contains 5 sections of code:

ParseState

Data type and helper functions for keeping track of source position

Result

Maybe-like data type including error string for failure

Parser a

Modified Parser library from Graham Huttons Programming in Haskell

Scanner

Configurable scanner/lexer taking a String and returning a list of Tokens

languageDefsParser

A set of Parsers specifically for the gmr file, and any other similar meta-languages

ParseState

data ParseState

Data type for storing the current parse position within the input code. Used in the Parser library.

Constructors

ParseState

Fields

line :: Int

column :: Int

pos :: Int

code :: String

rest :: String

showPos :: ParseState -> String

Generates a human readable parse position string, as shown below:

Line 1, Column 14
     function text#()
                  ^

parseState :: String -> ParseState

Builds a starting ParseState from an input string

inc :: ParseState -> ParseState

Increments a ParseState by looking at the next character, and adjusting the line/column numbers accordingly

incColumn :: ParseState -> ParseState

Increases the column number and global position of the ParseState, consumes a character

incLine :: ParseState -> ParseState

Increases the line number and global position, and resets the column number of a ParseState. Consumes a character.

nextChar :: ParseState -> Char

Gets the next character of the ParseState

hasNextChar :: ParseState -> Bool

Checks if the ParseState has another character.

fromPos :: String -> Int -> ParseState

Builds a parse state from an input string and global position, automatically working out the line and column numbers

lineStr :: ParseState -> String

Gets the current full line of the ParseState

Result

data Result a

Result data type, similar to Maybe (with definitions for Monad) but with an error string

Constructors

Result a

Error String

eitherToResult :: Either String a -> Result a

Converts an Either String a computation to a Result a

Parser

newtype Parser a

Parser data type, which transforms a ParseState to another ParseState, with the possibility to fail on the output

Constructors

P (ParseState -> (Maybe a, ParseState))

parseToResult :: Parser a -> String -> Result a

Takes a Parser and String, and returns a Result from the Parser. Fails if the Parser fails to return a value, or if the input is not fully used up. The ParseState will be used to give meaningful errors on failure.

parse :: Parser a -> ParseState -> (Maybe a, ParseState)

Simply runs a Parser on an input

item :: Parser Char

Gets the next character, or fails if there aren't any

peek :: Parser Char

Gets the next character without consuming it from the input, or fails if there isn't one

canPeek :: Parser Bool

Checks if the ParseState has a next input, as a Parser computation

tryPeek :: Parser (Maybe Char)

Attempts to peek, returning a Just Char if successful, or Nothing otherwise. This cannot fail.

getState :: Parser ParseState

Gets the current ParseState

mustFail :: Parser a -> Parser ()

Enforces a Parser to fail, essentially reversing the default failure behaviour.

sat :: (Char -> Bool) -> Parser Char

Ensures the next character satisfying a predicate, returning the character if it does.

digit :: Parser Char

Gets a digit

lower :: Parser Char

Gets a lower-case character

upper :: Parser Char

Gets an upper-case character

letter :: Parser Char

Gets a letter as defined by isAlpha

alphanum :: Parser Char

Gets an alphanumeric character

char :: Char -> Parser Char

Attempts to read a specific character

string :: String -> Parser String

Attempts to read a specific string

anyString :: [String] -> Parser String

Attempts to read any of a list of strings

ident :: Parser String

Attempts to read an identity, as defined by an alphanumeric string starting with a lower-case letter

upperIdent :: Parser String

Attempts to read an upper-identity, as defined by an alphanumeric string starting with a upper-case letter

nat :: Parser Int

Attempts to read a natural number

int :: Parser Int

Attempts to read an integer value

float :: Parser Float

Attempts to read a float value, which can also interpret integers

Scanner

data Scanner

Scanner data type containing the configuration for a scanner

Constructors

Scanner

Fields

separateCasedIdentifiers :: Bool

ignoreWhitespace :: Bool

ignoreComments :: Bool

operators :: [String]

keywords :: [String]

blockComment :: Maybe (String, String)

lineComment :: Maybe String

includeEOF :: Bool

parserMap :: [Parser TokenType] -> [Parser TokenType]

data Token

Token data type storing the token itself and the ParseState it was taken from

Constructors

Token ParseState TokenType

data TokenType

Various tokens for the scanner, including a TokenCustom constructor for any extras.

Constructors

TokenIdentifier String

TokenUpperIdentifier String

TokenOperator String

TokenStringLit String

TokenIntLit Int

TokenFloatLit Float

TokenKeyword String

TokenWhitespace Char

TokenBlockComment String

TokenLineComment String

TokenCustom String String

TokenOpenParen

TokenCloseParen

TokenOpenSquare

TokenCloseSquare

TokenOpenCurly

TokenCloseCurly

TokenEOF

scanner :: Scanner

Default definition of Scanner, for record modification

space :: Parser TokenType

Matches a single space, and wraps the result in a TokenWhiteSpace

tokenWithState :: Parser TokenType -> Parser Token

Takes a TokenType parser, gets the current ParseState, and returns an equivalent Token parser

asToken :: Scanner -> Parser TokenType -> Parser [Token]

Allows a token to be preceeded or proceeded by any amount of spaces, keeping them if the Scanner configuration says to.

formattedString :: Parser String

Reads a string literal, defined as any amount of tokens between two quotes, allowing them to be escaped with a if needed.

formattedStringRaw :: Parser String

Similar to formattedString, but keeps the quotes and escape characters in the outputted string

operator :: Scanner -> Parser TokenType

Creates a Parser for operators from the Scanner configuration

keyword :: Scanner -> Parser TokenType

Creates a Parser for keywords from the Scanner configuration

getCharParser :: Char -> TokenType -> Parser TokenType

Reads a character and returns a specific TokenType on read

nonAlphaNext :: Parser ()

Ensures the next character is not alpha

stringToken :: Parser TokenType

Parser for string literals, wrapped in TokenStringLit

intToken :: Parser TokenType

Parser for integers, wrapped in TokenIntLit

floatToken :: Parser TokenType

Parser for floats, wrapped in TokenFloatLit

lowerIdentifierToken :: Parser TokenType

Parser for lower identifiers, wrapped in TokenIdentifier

upperIdentifierToken :: Parser TokenType

Parser for upper identifiers, wrapped in TokenUpperIdentifier

identifierToken :: Parser TokenType

Parser for both identifiers and upper identifiers, both wrapped in TokenIdentifier

brackets :: Parser TokenType

Parsers for all of the 6 brackets: [] {} ()

getIdentifierTokenParser :: Bool -> Parser TokenType

Gets parsers for identifiers, either separated by upper and lower identifiers, or both treated as lower identifiers, based on the boolean input

commentParser :: String -> String -> (String -> TokenType) -> Parser TokenType

Parser for comments taking a starting and ending token, along with a constructor.

blockCommentParser :: Scanner -> Parser TokenType

Parser for block comments based on Scanner configuration

lineCommentParser :: Scanner -> Parser TokenType

Parser for line comments, using the Scanner for the starting token, and a new line for the ending token.

scanParser :: Scanner -> Parser [Token]

Creates a full scanner from a Scanner configuration

scan :: Scanner -> String -> Result [Token]

Runs a Scanner on a string, returning a Result of the token list

Language definition parsers

unpackCustom :: TokenType -> String

Unpacks the data value from a TokenCustom

codeBlock :: Parser TokenType

Haskell code block, defined as follows: { haskellCode }

directive :: Parser TokenType

Parser for directives of the form %mydirective

identifierPrime :: Parser TokenType

Haskell identifiers, defined as lower identifiers followed by zero or more 's

languageDefsParser :: Parser TokenType

All three above parsers combined into one parser