SemanticsCodeGenerator

Description

This file takes the SemanticsDef outputted by the parser after it has been validated, and generates the haskell code for the semantics checkers. It is responsible for setting up the environment for the reductions, and building all the error reporting mechanisms required.

Documentation

generateSemanticsCode

Arguments

:: String
Semantics module name

-> String
Parser module name

-> Maybe String
Optional additional imports

-> Maybe String
Optional "pre-code", a section of haskell code to be placed at the start of the file

-> Maybe String
Optional output "pre-code", a haskell String containing C code to be placed at the start of the output C file

-> SemanticsDef
The semantics definition from Semantics

-> String
The semantics checker code

Takes various meta information about the semantics checker, along with the SemanticsDef, and creates the full semantics checking file.