SemanticsValidator

Description

This file checks that any direct references to base types are valid, and performs converts the semantics rule types to actions to be taken by the code generator.

Documentation

validateSemantics :: SemanticsDef -> Result SemanticsDef

Entry point to the validator, handles all listed above.